:recycle: refactor: 예외 응답 DTO 패키지 구조 변경
